==> Customer Revenue Prediction - A Kaggle Competition/1. Preprocessing .ipynb <==
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

train_url = "C:/Users/kamyar.manshaei/Desktop/Kaggle/Google Competition/train.csv"
train = pd.read_csv(train_url)
test_url = "C:/Users/kamyar.manshaei/Desktop/Kaggle/Google Competition/test.csv"
test = pd.read_csv(test_url)

# <font color='red'>Size of the datasets

print("train size:")
print(train.shape)
print("test size:")
print(test.shape)

--------------------------------------------------------------------------------------------------------------------------------

# <font color='red'>Numeric and Categorical columns

numerics = train._get_numeric_data().columns.values.tolist()
categoricals = [col for col in train.columns.values if col not in numerics]

print('numeric columns:')
print(numerics)
print()
print('categorical columns:')
print(categoricals)

# <font color='red'>Removing useless columns

### <font color='blue'>Columns appeared only in train set

for col in train.columns.values:
    if col not in test.columns.values:
        print(col)
        print("Number of non-NaN values: ", len(train[col][train[col].isna() == False]))
        print()

train.drop('campaignCode',axis=1, inplace=True)
categoricals.remove('campaignCode')

### <font color='blue'>Removing columns containing only NaN

for col in train.columns.values:
    if len(train[col][train[col].isna() != False]) == train.shape[0]:
        print(col)

train.drop('targetingCriteria',axis=1, inplace=True)
test.drop('targetingCriteria',axis=1, inplace=True)
numerics.remove('targetingCriteria')

### <font color='blue'>Dropping columns with only one value

for col in train.columns.values:
    if len(train[col].unique()) == 1:
        print(col)
        train.drop(col,axis=1, inplace=True)
        test.drop(col,axis=1, inplace=True)
        if col in categoricals:
            categoricals.remove(col)
        if col in numerics:
            numerics.remove(col)

--------------------------------------------------------------------------------------------------------------------------------

# <font color='red'>Finding a unique ID

Just to check how many IDs we have:

[i for i in train.columns.values if 'Id' in i]

Here we find columns, in which more than 50% of the values are unique:

for col in train.columns.values:
    if (train[col].nunique() / train.shape[0]) > 0.5 :
        print(col)
        print (train[col].nunique() / train.shape[0])
        print()

for col in test.columns.values:
    if (test[col].nunique() / test.shape[0]) > 0.5 :
        print(col)
        print (test[col].nunique() / test.shape[0])
        print()

As we can see, there is no completely unique column. One probable reason could be presence of replicated rows. Therefore we check if there are replicated rows in the dataset:

train[train.duplicated()]

No replicated rows in the datset, but must be checked again after imputation of missing values

### <font color='blue'>Creating a unique ID

According to the description, *sessionId* is the unique ID and contains both *fullVistorId* and *visitId*, but as shown above, there are some duplicated values for *sessionId*, so we firstly check how the rows with identical *sessionIds* differ from each other. As there are many duplicated *sessionIds* and it takes a long time to show all of them, here only 20 duplicated pairs are shown (the number can be increased):

duplicated_ids = train["sessionId"][train["sessionId"].duplicated()].tolist()
from IPython.display import display 
j = 0
for i in duplicated_ids:
    dup = train[train["sessionId"] == i]
    columns = []
    for col in dup.columns.values:
        if(len(dup[col].unique()) > 1):
            columns.append(col)
    display(dup[columns])
    j += 1
    if(j == 20):
        break

It seems we can safely remove rows with duplicated ids, since it seems they only differ in time (and also the number of *hits* and *pageViews*, which apparently makes no differece, but must be checked later). The other differences are mostly caused by missing values. This duplication can be caused by changing the date during the session (the dates of these transactions are consecutive), e.g. transactions which have been started before 12:00 am and lasted until the next day. this claim can be checked later by converting the time column to a timestamp:

train.drop(train[train["sessionId"].duplicated()].index, inplace = True)
test.drop(test[test["sessionId"].duplicated()].index, inplace = True)

train.shape

As an alternative we can use the sum of *hits* and *pageViews* for duplicated *sessionIds*, instead of using one of them. This approach seems more logical, if hits and views indicate the total number of *hits* and *pageViews* occured during the transaction. Another important thing to be considered is that if the duplicated IDs represent one transaction, everywhere we have missing values only for one of them, we must replace them by the value of the other row.

### <font color='blue'>Removing useless IDs

As we noticed earlier, *sessionId* contain both *visitId* and *fullVisitorId*. Hence we can remove them without losing any information.:

train.drop('visitId',axis=1, inplace=True)
train.drop('fullVisitorId',axis=1, inplace=True)

test.drop('visitId',axis=1, inplace=True)
test.drop('fullVisitorId',axis=1, inplace=True)

numerics.remove('visitId')
categoricals.remove('fullVisitorId')

--------------------------------------------------------------------------------------------------------------------------------

# <font color='red'> Writing the final dataframe to .csv

train.to_csv('train1.csv', index=False, encoding='utf-8')
test.to_csv('test1.csv', index=False, encoding='utf-8')
