==> Customer Revenue Prediction - A Kaggle Competition/2. Filling missing values.ipynb <==
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

train_url = "C:/Users/kamyar.manshaei/Desktop/Kaggle/Google Competition/train1.csv"
train = pd.read_csv(train_url)
test_url = "C:/Users/kamyar.manshaei/Desktop/Kaggle/Google Competition/test1.csv"
test = pd.read_csv(test_url)

--------------------------------------------------------------------------------------------------------------------------------

# <font color='red'>Numeric and Categorical columns

numerics = train._get_numeric_data().columns.values.tolist()
categoricals = [col for col in train.columns.values if col not in numerics]

print('numeric columns:')
print(numerics)
print()
print('categorical columns:')
print(categoricals)

--------------------------------------------------------------------------------------------------------------------------------

# <font color='red'>Showing all missing values

d = {}
d_test = {}
types = []
for col in train.columns.values:
    d[col] = len(train[col][train[col].isna() == True])
    if col in numerics:
        types.append('numeric')
    else:
        types.append('categorical')
    #if len(train[col].isna())/train.shape[0] < 0.1:
        #train.drop(col,axis=1, inplace=True)
    if col in  test.columns.values:
        d_test[col] = len(test[col][test[col].isna() == True])
    else:
        d_test[col] = test.shape[0]
df = pd.DataFrame([], columns=['Column name', 'Missing values for train',"Missing values for test"])
df["Column name"] = d.keys()
df["Missing values for train"] = d.values()
df["Missing values for test"] = d_test.values()
df["Type"] = types
df["Train rate"] = df["Missing values for train"]/train.shape[0]
df["Test rate"] = df["Missing values for test"]/test.shape[0]
df.sort_values(by=['Missing values for train'], ascending=False)

------------------------

# <font color='red'>Overview of columns

for col in train.columns.values:
    print(col,":")
    print(train[col].describe())
    print()

---------------------------

# <font color='red'>Variables with only one value and NaN

for col in train.columns.values:
    if len(train[col].unique()) == 2 and train[col].isnull().values.any():
        print(col)
        print(train[col].unique())

It seems that for numeric columns *bounces* and *newVisits*, the missing values indicate 0, and for boolean columns *isTrueDirect* and *isVideoAd* the missing values indcate __False__ and __True__ respectively. Therefore instead of removing them, we impute their missing values:

train["bounces"].fillna(0,inplace=True)
train["newVisits"].fillna(0,inplace=True)
train["isTrueDirect"].fillna(False,inplace=True)
train["isVideoAd"].fillna(True,inplace=True)

test["bounces"].fillna(0,inplace=True)
test["newVisits"].fillna(0,inplace=True)
test["isTrueDirect"].fillna(False,inplace=True)
test["isVideoAd"].fillna(True,inplace=True)

Now it's better to convert boolean columns to numerics, 0 represents __False__ and 1 represents __True__:

train["isTrueDirect"][train["isTrueDirect"] == False] = 0.0
train["isTrueDirect"][train["isTrueDirect"] == True] = 1.0

train["isVideoAd"][train["isVideoAd"] == False] = 0.0
train["isVideoAd"][train["isVideoAd"] == True] = 1.0

test["isTrueDirect"][test["isTrueDirect"] == False] = 0.0
test["isTrueDirect"][test["isTrueDirect"] == True] = 1.0

test["isVideoAd"][test["isVideoAd"] == False] = 0.0
test["isVideoAd"][test["isVideoAd"] == True] = 1.0

It's worth to check if there are other boolean columns:

for col in train.columns.values:
    if False in train[col].unique() and len(train[col].unique()) == 2 :
        print(col)
        print(train[col].unique())
        print()

Now we convert *isMobile* to numeric:

train["isMobile"][train["isMobile"] == False] = 0.0
train["isMobile"][train["isMobile"] == True] = 1.0

test["isMobile"][test["isMobile"] == False] = 0.0
test["isMobile"][test["isMobile"] == True] = 1.0

----------------------------------------------------------------------

# <font color='red'>A brief look into the target

*transactionRevenue* is the target and missing values probably indicate 0 revenue

print("Number of non-NaN values: ", len(train["transactionRevenue"][train["transactionRevenue"].isna() == False]))

### <font color='blue'>Converting target missings to 0</font>

train["transactionRevenue"].fillna(0,inplace=True)

----------------------------------------------------------------------

# <font color='red'>Imputing missing values

Let's see how many columns still contain missing values:

d = {}
d_test = {}
types = []
for col in train.columns.values:
    d[col] = len(train[col][train[col].isna() == True])
    if col in numerics:
        types.append('numeric')
    else:
        types.append('categorical')
    #if len(train[col].isna())/train.shape[0] < 0.1:
        #train.drop(col,axis=1, inplace=True)
    if col in  test.columns.values:
        d_test[col] = len(test[col][test[col].isna() == True])
    else:
        d_test[col] = test.shape[0]
df = pd.DataFrame([], columns=['Column name', 'Missing values for train',"Missing values for test"])
df["Column name"] = d.keys()
df["Missing values for train"] = d.values()
df["Missing values for test"] = d_test.values()
df["Type"] = types
df["Train rate"] = df["Missing values for train"]/train.shape[0]
df["Test rate"] = df["Missing values for test"]/test.shape[0]
df[df['Missing values for train'] > 0].sort_values(by=['Missing values for train'], ascending=False)

for col in df["Column name"][df['Missing values for train'] > 0]:
    print(col)
    print(train[col].describe())
    print()

For those categorical variables which have less than 70% missing values, missings are replaced with mode (categoricals) and median (numerics):

train["pageviews"].fillna(train["pageviews"].median(),inplace=True)
train["keyword"].fillna(train["keyword"].mode()[0],inplace=True)
train["referralPath"].fillna(train["referralPath"].mode()[0],inplace=True)

test["pageviews"].fillna(test["pageviews"].median(),inplace=True)
test["keyword"].fillna(test["keyword"].mode()[0],inplace=True)
test["referralPath"].fillna(test["referralPath"].mode()[0],inplace=True)

The other variables contain more than 97% missing values, so they can be removed:

for col in train.columns.values:
    if len(train[col][train[col].isna()==True])/train.shape[0] > 0.90:
        print(col)
        train.drop(col,axis=1, inplace=True)
        test.drop(col,axis=1, inplace=True)
        if col in categoricals:
            categoricals.remove(col)
        if col in numerics:
            numerics.remove(col)

Checking if there are still missing values in the dataset:

train.isnull().values.any()

# <font color='red'> Writing the final dataframe to .csv

train.to_csv('train2.csv', index=False, encoding='utf-8')
test.to_csv('test2.csv', index=False, encoding='utf-8')
