==> Customer Revenue Prediction - A Kaggle Competition/3. Making valid types.ipynb <==
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

train_url = "C:/Users/kamyar.manshaei/Desktop/Kaggle/Google Competition/train2.csv"
train = pd.read_csv(train_url)
test_url = "C:/Users/kamyar.manshaei/Desktop/Kaggle/Google Competition/test2.csv"
test = pd.read_csv(test_url)

# <font color='red'>Numeric and Categorical columns

numerics = train._get_numeric_data().columns.values.tolist()
categoricals = [col for col in train.columns.values if col not in numerics]

print('numeric columns:')
print(numerics)
print()
print('categorical columns:')
print(categoricals)

# <font color='red'>Making valid types for boolean variables

Before going further, we need to check if all columns are correctly assigned to their categories based on their types:

for col in categoricals:
    print(col)
    print(train[col].describe())
    print()

For categoricals, everything looks fine so far. <font size="+2">&#128077;

Now we will have the same procedure this time for numerics:

for col in numerics:
    print(col)
    print(train[col].describe())
    print()

Some of the variables in numerics are still boolean and must be converted to integer:

train["isMobile"][train["isMobile"] == True] = 1
train["isMobile"][train["isMobile"] == False] = 0
train["isVideoAd"][train["isVideoAd"] == True] = 1
train["isVideoAd"][train["isVideoAd"] == False] = 0
train["isTrueDirect"][train["isTrueDirect"] == True] = 1
train["isTrueDirect"][train["isTrueDirect"] == False] = 0

test["isMobile"][test["isMobile"] == True] = 1
test["isMobile"][test["isMobile"] == False] = 0
test["isVideoAd"][test["isVideoAd"] == True] = 1
test["isVideoAd"][test["isVideoAd"] == False] = 0
test["isTrueDirect"][test["isTrueDirect"] == True] = 1
test["isTrueDirect"][test["isTrueDirect"] == False] = 0

-------------------------------------------------------

# <font color='red'>Dealing with times and dates

Columns *date* and *visitStartTime* ,which represent date and time, must be handled somehow.

### <font color='blue'>Date

At first we split *date* into day, month and year:

train["date"] = train["date"].apply(str)
train["year"] = train["date"].str.slice(0, 4)
train["month"] = train["date"].str.slice(4, 6)
train["day"] = train["date"].str.slice(6, 8)

test["date"] = test["date"].apply(str)
test["year"] = test["date"].str.slice(0, 4)
test["month"] = test["date"].str.slice(4, 6)
test["day"] = test["date"].str.slice(6, 8)

Now we can remove *date* from the dataset:

train.drop('date',axis=1, inplace=True)
test.drop('date',axis=1, inplace=True)

numerics.remove('date')
categoricals.append('year')
categoricals.append('month')
categoricals.append('day')

### <font color='blue'>Time 

Now we are going to deal with *visitStartTime*. This variable will be splitted into two variables:
<ol>
1) *weekday*, which indicates the day of the week<br>
2) *startHour*, which indicates the hour, at which the transaction was started
</ol>

The next cell is only to import the date library and show how it works:

from datetime import datetime
print(datetime.utcfromtimestamp(train["visitStartTime"][0]).strftime('%Y-%m-%d %H:%M:%S'))

Now the weekday is extracted from *visitStartTime* and attached to the dataframe:

train["weekday"] = train["visitStartTime"].apply(datetime.utcfromtimestamp).apply(datetime.weekday)
test["weekday"] = test["visitStartTime"].apply(datetime.utcfromtimestamp).apply(datetime.weekday)

And the hour:

train["startHour"] = train["visitStartTime"].apply(datetime.utcfromtimestamp).dt.hour
test["startHour"] = test["visitStartTime"].apply(datetime.utcfromtimestamp).dt.hour

Now we can remove *visitStartTime* from the dataset and add our new features:

train.drop('visitStartTime',axis=1, inplace=True)
test.drop('visitStartTime',axis=1, inplace=True)
numerics.remove('visitStartTime')

numerics.append('weekday')
numerics.append('startHour')

<pre> <font color='green' size="+1"> __We could have extracted all the new columns from one of the variables *date* and *visitStartTime*!__ &#x1F60E;

# <font color='red'> Writing the final dataframe to .csv

train.to_csv('train3.csv', index=False, encoding='utf-8')
test.to_csv('test3.csv', index=False, encoding='utf-8')
