==> Customer Revenue Prediction - A Kaggle Competition/5. Primitive dimensionality reduction.ipynb <==
import pandas as pd
import numpy as np
import warnings
from collections import Counter
warnings.filterwarnings('ignore')

train_url = "C:/Users/kamyar.manshaei/Desktop/Kaggle/Google Competition/train4.csv"
train = pd.read_csv(train_url)
test_url = "C:/Users/kamyar.manshaei/Desktop/Kaggle/Google Competition/test4.csv"
test = pd.read_csv(test_url)

--------------------------------------

# <font color='red'>Numeric and Categorical columns

numerics = train._get_numeric_data().columns.values.tolist()
categoricals = [col for col in train.columns.values if col not in numerics]

print('numeric columns:')
print(numerics)
print()
print('categorical columns:')
print(categoricals)

### <font color='blue'> Removal of less-informative columns

In order to prevent curse of dimensionality, we can remove columns with a large number of categories, but low information:

train.drop(['continent','region','metro'], axis=1, inplace=True)
test.drop(['continent','region','metro'], axis=1, inplace=True)
categoricals.remove('continent')
categoricals.remove('region')
categoricals.remove('metro')

### <font color='blue'> Replacing categories with frequencies

As some columns contain values occured only in one of train or test sets, we set them to _others_: (E.g. _**country**_, _**source**_, _**city**_ and ...)

categoricals.remove('sessionId')
for col in categoricals:
    print(col)
    for val in train[col].unique().tolist():
        if val not in test[col].unique().tolist():
            train[col][train[col] == val] = 'others'

    for val in test[col].unique().tolist():
        if val not in train[col].unique().tolist():
            test[col][test[col] == val] = 'others'

All categorical columns can be converted to integer by replacing their category with the frequency of the category: <font color='green'> __(WITHOUT THIS, WE WILL GET A MEMORY ERROR DURING PCA and XGBoost)__

for col in categoricals:
    print(col)
    train['temp'] = float('NaN')
    test['temp'] = float('NaN')
    for value in train[col].unique().tolist():
        train['temp'][train[col] == value] = Counter(train[col].tolist())[value]
        test['temp'][test[col] == value] = Counter(test[col].tolist())[value]
    
    train[col] = train['temp']
    test[col] = test['temp']
    train.drop('temp',axis=1, inplace=True)
    test.drop('temp',axis=1, inplace=True)

------------------------------------

# <font color='red'> Writing the final dataframe to .csv

train.to_csv('train5.csv', index=False, encoding='utf-8')
test.to_csv('test5.csv', index=False, encoding='utf-8')
